Including model for shakespeare and paths for flower baseline
